requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

boto3[dynamodb]==1.38.27
botocore==1.38.27
dacite==1.9.2
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.13.0
six==1.17.0
urllib3==2.4.0; python_version >= "3.10"

